Split SourceCodeRepoFactoryTest.java into two files
