+++ dockstore-webservice/src/test/java/io/dockstore/webservice/helpers/SourceCodeRepoFactoryTest.java
@@ -179,55 +179,4 @@ public class SourceCodeRepoFactoryTest {
 
     }
 
-    /**
-     * Tests that parsing a BitBucket URL works
-     */
-    @Test
-    public void testBitBucketUrlParsing() {
-        BitBucketSourceCodeRepo repo = new BitBucketSourceCodeRepo("fakeUser", "fakeToken");
-        final BioWorkflow entry = new BioWorkflow();
-
-        /* Test good URLs */
-        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui.git");
-        String bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", "dockstore/dockstore-ui", bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", "dockstore/dockstore-ui", bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore-cow/goat.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", "dockstore-cow/goat", bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore.dot/goat.bat.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", "dockstore.dot/goat.bat", bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore.dot/goat.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", "dockstore.dot/goat", bitBucketId);
-
-        /* Test bad URLs */
-        entry.setGitUrl("git@bitbucket.org/dockstore/dockstore-ui.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", null, bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore:dockstore-ui.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", null, bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:/dockstore-ui.git");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", null, bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", null, bitBucketId);
-
-        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui");
-        bitBucketId = repo.getRepositoryId(entry);
-        assertEquals("Bitbucket ID parse check", null, bitBucketId);
-
-    }
 }
